Reclassify as structured vs. image Clean up copy paste errors
